chore: foundation — PHPStan level 5 in CI, Jest for diff.js, drop dead code
Sprint 1 of the post-audit cleanup. No behaviour change for the user; the
focus is on getting type-checker + JS-side tests landing in CI so future
PRs catch regressions earlier. Started from 52 PHPStan errors at level 5
and went to zero.

PHPStan stack (level 5, runs over `src/`):
- phpstan/phpstan ^2.2
- szepeviktor/phpstan-wordpress ^2.0
- php-stubs/wp-cli-stubs ^2.12

Type / shape fixes:
- ToolProviderInterface::getTools() now returns
  list<array{name, description, input_schema}> matching what every
  provider actually emits (was array<string, ...>, which mismatched 6
  callers and silently allowed the wrong shape through).
- Removed the unused $plugin constructor arg + property from
  ChatEndpoint and ConversationEndpoint.
- Plugin::getInstance() now returns self instead of unsafe new static().
- ProviderRegistry::$providers PHPDoc shape extended to cover
  connector_id / label / base_url / default (was previously declared as
  a closed shape that excluded them); resolve() return shape gained the
  `tier` field it always carried; getModelsForFrontend() shape gained
  `pricing`.
- AbilitiesToolProvider::$toolCache PHPDoc tightened to the real
  list-of-records shape.

Dropped redundant defensive checks where PHPStan could prove the values
were already narrowed:
- is_array() on already-typed arrays
- is_wp_error() on int returns from wp_insert_post /
  wp_insert_attachment (passing $wp_error=true so failures actually
  surface as WP_Error rather than 0)
- null-arm of the $body comparison in McpClient (it's non-empty-string
  at that point)
- false-arm of wp_salt() in Encrypt
- dead Throwable catch around the Laravel Log facade call (the
  class_exists check already gates Acorn-not-booted)
- redundant array_values() on an already-list result in ToolRestrictor

Real bugs surfaced + fixed:
- MessageLoop now (int)-casts incoming usage tokens — a string-shaped
  token field would have silently accumulated to 0 and broken the daily
  budget tracking.
- Plugin.php's bare `catch (\Throwable)` around Polylang's
  get_languages_list() now logs the exception message via error_log()
  instead of swallowing it, so a real misconfiguration surfaces in
  debug.log rather than degrading to "no voice languages" with no
  signal.
- wp_insert_post / wp_insert_attachment now pass $wp_error=true so
  failures surface as WP_Error rather than 0.
- Gemini provider: rebuilt the JSON content-decoding branch so PHPStan
  can follow it; fixed the
  `'image/'.pathinfo(...) ?: 'jpeg'` precedence bug that made the jpeg
  fallback unreachable.

JS unit tests (wp-scripts test-unit-js → Jest):
- New resources/scripts/editor/__tests__/diff.test.js covers all four
  exports of editor/diff.js: diffLines, diffWords, collapseUnchanged,
  pairModifiedLines. 17 cases including the end-to-end shape used by
  the chat tool-call diff card.

CI:
- New `composer stan` script.
- Lint job in .github/workflows/test.yml now also runs `composer stan`
  and `npm run test:unit` so PHPStan + Jest land alongside Pint.

+++ src/Llm/MessageLoop.php
@@ -74,12 +74,17 @@ class MessageLoop
 
         // Track token usage across iterations. All providers now emit
         // unified field names: cache_read_tokens / cache_write_tokens.
+        // Explicit (int) casts: a provider that ever emits a string ("1024")
+        // instead of an int would silently accumulate to 0 with the bare
+        // `+= ($data[...] ?? 0)`, breaking budget tracking. Cast at the
+        // boundary so accounting matches the type the rest of the loop
+        // expects.
         $wrappedOnEvent = function (string $type, array $data) use ($onEvent) {
             if ($type === 'usage') {
-                $this->inputTokens += $data['input_tokens'] ?? 0;
-                $this->outputTokens += $data['output_tokens'] ?? 0;
-                $this->cacheCreationTokens += $data['cache_write_tokens'] ?? 0;
-                $this->cacheReadTokens += $data['cache_read_tokens'] ?? 0;
+                $this->inputTokens += (int) ($data['input_tokens'] ?? 0);
+                $this->outputTokens += (int) ($data['output_tokens'] ?? 0);
+                $this->cacheCreationTokens += (int) ($data['cache_write_tokens'] ?? 0);
+                $this->cacheReadTokens += (int) ($data['cache_read_tokens'] ?? 0);
             }
             $onEvent($type, $data);
         };
